# backend/openapi.yaml
openapi: 3.0.3
info:
  title: PasteHub Backend API
  version: 1.0.0
servers:
  - url: http://localhost:5001
paths:
...
  /api/v1/pastes/file:
    post:
      summary: Upload file paste (multipart/form-data)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                visibility:
                  type: string
                  enum: [public, unlisted, private]
                expires_in:
                  type: string
                  enum: [10m, 1h, 1d, 1w, 1m, never]
                password:
                  type: string
                tags:
                  type: string
      responses:
        '201': { description: Created }
...
  /api/v1/pastes/{slug}:
    get:
      summary: Read paste with metadata and content
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        '200': { description: Paste found }
        '404': { description: Not found }
    patch:
      summary: Update paste
      responses:
        '200': { description: Updated }
    delete:
      summary: Delete paste
      responses:
        '200': { description: Deleted }
...